feat: Adding support for `--filter` in `stack generate`
