+++ config/stack.go
@@ -3,7 +3,6 @@ package config
 import (
 	"context"
 	"fmt"
-	"io/fs"
 	"os"
 	"path/filepath"
 	"sort"
@@ -16,7 +15,6 @@ import (
 	"github.com/gruntwork-io/terragrunt/internal/ctyhelper"
 	"github.com/gruntwork-io/terragrunt/internal/worker"
 
-	"github.com/gruntwork-io/terragrunt/internal/experiment"
 	"github.com/hashicorp/go-getter/v2"
 
 	"github.com/hashicorp/hcl/v2/hclsyntax"
@@ -35,7 +33,6 @@ const (
 	StackDir          = ".terragrunt-stack"
 	valuesFile        = "terragrunt.values.hcl"
 	manifestName      = ".terragrunt-stack-manifest"
-	defaultStackFile  = "terragrunt.stack.hcl"
 	unitDirPerm       = 0755
 	valueFilePerm     = 0644
 	generationMaxPath = 1024
@@ -75,457 +72,10 @@ type Stack struct {
 	Path         string     `hcl:"path,attr"`
 }
 
-// StackNode represents a stack file in the file system.
-// The parent is the node that generates the current node,
-// and children are the nodes that are generated by the current node.
-type StackNode struct {
-	Parent   *StackNode
-	FilePath string
-	Children []*StackNode
-	Level    int
-}
-
-// NewStackNode creates a new stack node.
-func NewStackNode(filePath string) *StackNode {
-	return &StackNode{
-		FilePath: filePath,
-		Level:    -1,
-		Children: make([]*StackNode, 0),
-	}
-}
-
-// GenerateStacks generates the stack files using topological ordering to prevent race conditions.
-// Stack files are generated level by level, ensuring parent stacks complete before their children.
-func GenerateStacks(ctx context.Context, l log.Logger, opts *options.TerragruntOptions) error {
-	foundFiles, err := listStackFiles(l, opts, opts.WorkingDir)
-	if err != nil {
-		return errors.Errorf("Failed to list stack files in %s %w", opts.WorkingDir, err)
-	}
-
-	if len(foundFiles) == 0 {
-		if opts.StackAction == "generate" {
-			l.Warnf("No stack files found in %s Nothing to generate.", opts.WorkingDir)
-		}
-
-		return nil
-	}
-
-	generatedFiles := make(map[string]bool)
-
-	stackTrees := BuildStackTopology(l, foundFiles, opts.WorkingDir)
-
-	const maxLevel = 1024
-	for level := range maxLevel {
-		if level == maxLevel-1 {
-			return errors.Errorf("Cycle detected: maximum level (%d) exceeded", maxLevel)
-		}
-
-		levelNodes := getNodesAtLevel(stackTrees, level)
-		if len(levelNodes) == 0 {
-			break
-		}
-
-		if err := generateLevel(ctx, l, opts, level, levelNodes, generatedFiles); err != nil {
-			return err
-		}
-
-		if err := discoverAndAddNewNodes(l, opts, stackTrees, generatedFiles, level+1); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-// generateLevel handles the concurrent generation of all stack files at a given level.
-func generateLevel(ctx context.Context, l log.Logger, opts *options.TerragruntOptions, level int, levelNodes []*StackNode, generatedFiles map[string]bool) error {
-	l.Debugf("Generating stack level %d with %d files", level, len(levelNodes))
-
-	wp := worker.NewWorkerPool(opts.Parallelism)
-	defer wp.Stop()
-
-	for _, node := range levelNodes {
-		if generatedFiles[node.FilePath] {
-			continue
-		}
-
-		generatedFiles[node.FilePath] = true
-
-		// Before attempting to generate the stack file, we need to double-check that the file exists.
-		// Generation at a higher level might have resulted in this file being removed.
-		if !util.FileExists(node.FilePath) {
-			continue
-		}
-
-		wp.Submit(func() error {
-			return generateStackFile(ctx, l, opts, wp, node.FilePath)
-		})
-	}
-
-	return wp.Wait()
-}
-
-// discoverAndAddNewNodes discovers new stack files and adds them to the dependency graph.
-func discoverAndAddNewNodes(l log.Logger, opts *options.TerragruntOptions, dependencyGraph map[string]*StackNode, generatedFiles map[string]bool, minLevel int) error {
-	newFiles, listErr := listStackFiles(l, opts, opts.WorkingDir)
-	if listErr != nil {
-		return errors.Errorf("Failed to list stack files after level %d: %w", minLevel-1, listErr)
-	}
-
-	addNewNodesToGraph(l, dependencyGraph, newFiles, generatedFiles, opts.WorkingDir)
-
-	return nil
-}
-
-// BuildStackTopology creates a topological tree based on directory hierarchy.
-func BuildStackTopology(l log.Logger, stackFiles []string, workingDir string) map[string]*StackNode {
-	nodes := make(map[string]*StackNode)
-
-	for _, file := range stackFiles {
-		nodes[file] = NewStackNode(file)
-	}
-
-	for _, node := range nodes {
-		assignNodeLevel(l, node, nodes, workingDir)
-	}
-
-	return nodes
-}
-
-// assignNodeLevel recursively assigns levels to nodes based on directory depth.
-func assignNodeLevel(l log.Logger, node *StackNode, allNodes map[string]*StackNode, workingDir string) int {
-	if node.Level != -1 {
-		return node.Level
-	}
-
-	nodeDir := filepath.Dir(node.FilePath)
-	parentPath := findParentStackFile(nodeDir, allNodes, workingDir)
-
-	if parentPath == "" {
-		node.Level = 0
-
-		return node.Level
-	}
-
-	parent := allNodes[parentPath]
-	if parent == nil {
-		node.Level = 0
-
-		return node.Level
-	}
-
-	parentLevel := assignNodeLevel(l, parent, allNodes, workingDir)
-	node.Level = parentLevel + 1
-	node.Parent = parent
-	parent.Children = append(parent.Children, node)
-
-	l.Debugf("Stack %s (level %d) is child of %s (level %d)", node.FilePath, node.Level, parent.FilePath, parent.Level)
-
-	return node.Level
-}
-
-// findParentStackFile finds the parent stack file for a given directory.
-func findParentStackFile(childDir string, allNodes map[string]*StackNode, workingDir string) string {
-	currentDir := childDir
-
-	for {
-		parentDir := filepath.Dir(currentDir)
-		if parentDir == currentDir {
-			break
-		}
-
-		if parentDir == workingDir {
-			potentialParent := filepath.Join(workingDir, defaultStackFile)
-			if _, exists := allNodes[potentialParent]; exists {
-				return potentialParent
-			}
-
-			break
-		}
-
-		potentialParent := filepath.Join(parentDir, defaultStackFile)
-		if _, exists := allNodes[potentialParent]; exists {
-			return potentialParent
-		}
-
-		currentDir = parentDir
-	}
-
-	return ""
-}
-
-// getNodesAtLevel returns all nodes at a specific level.
-func getNodesAtLevel(nodes map[string]*StackNode, level int) []*StackNode {
-	var levelNodes []*StackNode
-
-	for _, node := range nodes {
-		if node.Level == level {
-			levelNodes = append(levelNodes, node)
-		}
-	}
-
-	return levelNodes
-}
-
-// addNewNodesToGraph adds newly discovered stack files to the dependency graph.
-func addNewNodesToGraph(
-	l log.Logger,
-	existingNodes map[string]*StackNode,
-	allFiles []string,
-	generatedFiles map[string]bool,
-	workingDir string,
-) {
-	newFiles := make([]string, 0)
-
-	for _, file := range allFiles {
-		if _, exists := existingNodes[file]; !exists && !generatedFiles[file] {
-			newFiles = append(newFiles, file)
-		}
-	}
-
-	if len(newFiles) == 0 {
-		return
-	}
-
-	l.Debugf("Adding %d new stack files to topology graph", len(newFiles))
-
-	for _, file := range newFiles {
-		existingNodes[file] = NewStackNode(file)
-	}
-
-	for _, file := range newFiles {
-		node := existingNodes[file]
-		assignNodeLevel(l, node, existingNodes, workingDir)
-	}
-}
-
-// StackOutput collects and returns the OpenTofu/Terraform output values for all declared units in a stack hierarchy.
-//
-// This function is a central component of Terragrunt's stack output system, providing a mechanism to
-// aggregate and organize outputs from multiple deployments in a hierarchical structure. It's particularly
-// useful when working with complex infrastructure composed of multiple interconnected OpenTofu/Terraform units.
-//
-// The function performs several key operations:
-//
-//  1. Discovers all stack definition files (terragrunt.stack.hcl) in the working directory and its subdirectories.
-//  2. For each stack file, parses the configuration and extracts the declared stacks and units.
-//  3. For each unit, reads its OpenTofu/Terraform outputs from the corresponding directory within .terragrunt-stack.
-//  4. Constructs a hierarchical map of outputs by organizing units according to their position in the stack hierarchy.
-//     Units are keyed using dot notation that reflects the stack path (e.g., "parent.child.unit").
-//  5. Orders stack names from the highest level (shortest path) to deepest nested (longest path).
-//  6. Nests the flat output map into a hierarchical structure and converts it to a cty.Value object.
-//
-// The returned cty.Value object contains a structured representation of all outputs, preserving the
-// nested relationship between stacks and units. This makes it easy to access outputs from specific
-// parts of the infrastructure while maintaining awareness of the overall architecture.
-//
-// For telemetry and debugging purposes, the function logs various events at the debug level, including
-// when outputs are added for specific units and stack keys.
-//
-// Parameters:
-//   - ctx: Context for the operation, which may include telemetry collection.
-//   - opts: TerragruntOptions containing configuration settings and the working directory path.
-//
-// Returns:
-//   - cty.Value: A hierarchical object containing all outputs from the stack units, organized by stack path.
-//   - error: An error if any operation fails during discovery, parsing, output collection, or conversion.
-//
-// Errors can occur during stack file listing, value reading, stack config parsing, output reading,
-// or when converting the final output structure to cty.Value format.
-func StackOutput(ctx context.Context, l log.Logger, opts *options.TerragruntOptions) (cty.Value, error) {
-	l.Debugf("Generating output from %s", opts.WorkingDir)
-
-	foundFiles, err := listStackFiles(l, opts, opts.WorkingDir)
-	if err != nil {
-		return cty.NilVal, errors.Errorf("Failed to list stack files in %s: %w", opts.WorkingDir, err)
-	}
-
-	if len(foundFiles) == 0 {
-		l.Warnf("No stack files found in %s Nothing to generate.", opts.WorkingDir)
-		return cty.NilVal, nil
-	}
-
-	outputs := make(map[string]map[string]cty.Value)
-	declaredStacks := make(map[string]string)
-	declaredUnits := make(map[string]*Unit)
-
-	// save parsed stacks
-	parsedStackFiles := make(map[string]*StackConfig, len(foundFiles))
-
-	for _, path := range foundFiles {
-		dir := filepath.Dir(path)
-
-		values, valuesErr := ReadValues(ctx, l, opts, dir)
-		if valuesErr != nil {
-			return cty.NilVal, errors.Errorf("Failed to read values from %s: %w", dir, valuesErr)
-		}
-
-		stackFile, stackErr := ReadStackConfigFile(ctx, l, opts, path, values)
-		if stackErr != nil {
-			return cty.NilVal, errors.Errorf("Failed to read stack file %s: %w", path, stackErr)
-		}
-
-		parsedStackFiles[path] = stackFile
-
-		targetDir := filepath.Join(dir, StackDir)
-
-		for _, stack := range stackFile.Stacks {
-			declaredStacks[filepath.Join(targetDir, stack.Path)] = stack.Name
-			l.Debugf("Registered stack %s at path %s", stack.Name, filepath.Join(targetDir, stack.Path))
-		}
-
-		for _, unit := range stackFile.Units {
-			unitDir := getUnitDir(dir, unit)
-
-			var output map[string]cty.Value
-
-			telemetryErr := telemetry.TelemeterFromContext(ctx).Collect(ctx, "unit_output", map[string]any{
-				"unit_name":   unit.Name,
-				"unit_source": unit.Source,
-				"unit_path":   unit.Path,
-			}, func(ctx context.Context) error {
-				var outputErr error
-
-				output, outputErr = unit.ReadOutputs(ctx, l, opts, unitDir)
-
-				return outputErr
-			})
-			if telemetryErr != nil {
-				return cty.NilVal, errors.New(telemetryErr)
-			}
-
-			key := filepath.Join(targetDir, unit.Path)
-			declaredUnits[key] = unit
-			outputs[key] = output
-
-			l.Debugf("Added output for %s", key)
-		}
-	}
-
-	unitOutputs := make(map[string]map[string]cty.Value)
-
-	// Build stack list separated by stacks, find all nested stacks, and build a dotted path. If no stack is found, use the unit name.
-	for path, unit := range declaredUnits {
-		output, found := outputs[path]
-		if !found {
-			l.Debugf("No output found for %s", path)
-			continue
-		}
-
-		// Implement more logic to find all stacks in which the path is located
-		stackNames := []string{}
-		nameToPath := make(map[string]string) // Map to track which path each stack name came from
-
-		for stackPath, stackName := range declaredStacks {
-			if strings.Contains(path, stackPath) {
-				stackNames = append(stackNames, stackName)
-				nameToPath[stackName] = stackPath
-			}
-		}
-
-		// Sort stackNames based on the length of stackPath to ensure correct order
-		stackNamesSorted := make([]string, len(stackNames))
-		copy(stackNamesSorted, stackNames)
-
-		for i := range stackNamesSorted {
-			for j := i + 1; j < len(stackNamesSorted); j++ {
-				// Compare lengths of the actual paths from the nameToPath map, not the declaredStacks lookup
-				if len(nameToPath[stackNamesSorted[i]]) < len(nameToPath[stackNamesSorted[j]]) {
-					stackNamesSorted[i], stackNamesSorted[j] = stackNamesSorted[j], stackNamesSorted[i]
-				}
-			}
-		}
-
-		stackKey := unit.Name
-		if len(stackNamesSorted) > 0 {
-			stackKey = strings.Join(stackNamesSorted, ".") + "." + unit.Name
-		}
-
-		unitOutputs[stackKey] = output
-
-		l.Debugf("Added output for stack key %s", stackKey)
-	}
-
-	// Convert finalMap into a cty.ObjectVal
-	result := make(map[string]cty.Value)
-
-	nestedOutputs, err := nestUnitOutputs(unitOutputs)
-	if err != nil {
-		return cty.NilVal, errors.Errorf("Failed to nest unit outputs: %w", err)
-	}
-
-	ctyResult, err := goTypeToCty(nestedOutputs)
-	if err != nil {
-		return cty.NilVal, errors.Errorf("Failed to convert unit output to cty value: %s %w", result, err)
-	}
-
-	return ctyResult, nil
-}
-
-// nestUnitOutputs transforms a flat map of unit outputs into a nested hierarchical structure.
-//
-// This function is a critical part of Terragrunt/Opentofu's stack output system, converting flat key-value pairs
-// with dot notation into a proper nested object hierarchy. It processes each flattened key (e.g., "parent.child.unit")
-// by splitting it into path segments and recursively building the corresponding nested structure.
-//
-// The algorithm works as follows:
-//  1. For each entry in the flat map, split its key by dots to get the path segments
-//  2. Iteratively traverse the nested structure, creating intermediate maps as needed
-//  3. When reaching the final path segment, convert the map of cty.Values to a Go interface{}
-//     representation and store it at that location
-//  4. Continue until all flat entries have been properly nested
-//
-// This approach preserves the hierarchical relationship between stacks and units while making
-// the data structure easier to navigate and query programmatically.
-//
-// Parameters:
-//   - flat: A map where keys are dot-separated paths (e.g., "parent.child.unit") and values are
-//     maps of cty.Value representing the OpenTofu/Terraform outputs for each unit
-//
-// Returns:
-//   - map[string]any: A nested map structure reflecting the hierarchy implied by the dot notation
-//   - error: An error if conversion fails, particularly when building the nested structure
-//
-// Errors can occur during cty.Value conversion or when attempting to traverse the nested structure
-// if the path contains contradictory type information (e.g., a path segment is both a leaf and a branch).
-func nestUnitOutputs(flat map[string]map[string]cty.Value) (map[string]any, error) {
-	nested := make(map[string]any)
-
-	for flatKey, value := range flat {
-		parts := strings.Split(flatKey, ".")
-		current := nested
-
-		for i, part := range parts {
-			if i == len(parts)-1 {
-				ctyValue, err := convertValuesMapToCtyVal(value)
-				if err != nil {
-					return nil, errors.Errorf("Failed to convert unit output to cty value: %s %w", flatKey, err)
-				}
-
-				current[part] = ctyValue
-			} else {
-				if _, exists := current[part]; !exists { // Traverse or create next level
-					current[part] = make(map[string]any)
-				}
-
-				var ok bool
-
-				current, ok = current[part].(map[string]any)
-
-				if !ok {
-					return nil, errors.Errorf("Failed to traverse unit output: %v %s", flat, part)
-				}
-			}
-		}
-	}
-
-	return nested, nil
-}
-
-// generateStackFile generates the Terragrunt stack configuration from the given stackFilePath,
+// GenerateStackFile generates the Terragrunt stack configuration from the given stackFilePath,
 // reads necessary values, and generates units and stacks in the target directory.
 // It handles the creation of required directories and returns any errors encountered.
-func generateStackFile(ctx context.Context, l log.Logger, opts *options.TerragruntOptions, pool *worker.Pool, stackFilePath string) error {
+func GenerateStackFile(ctx context.Context, l log.Logger, opts *options.TerragruntOptions, pool *worker.Pool, stackFilePath string) error {
 	stackSourceDir := filepath.Dir(stackFilePath)
 
 	values, err := ReadValues(ctx, l, opts, stackSourceDir)
@@ -723,7 +273,7 @@ func generateComponent(ctx context.Context, l log.Logger, opts *options.Terragru
 		expectedFile := DefaultTerragruntConfigPath
 
 		if cmp.kind == stackKind {
-			expectedFile = defaultStackFile
+			expectedFile = DefaultStackFile
 		}
 
 		if err := validateTargetDir(kindStr, cmp.name, dest, expectedFile); err != nil {
@@ -1073,7 +623,7 @@ func processLocals(l log.Logger, parser *ParsingContext, opts *options.Terragrun
 		}
 	}
 
-	localsAsCtyVal, err := convertValuesMapToCtyVal(evaluatedLocals)
+	localsAsCtyVal, err := ConvertValuesMapToCtyVal(evaluatedLocals)
 	if err != nil {
 		return errors.New(err)
 	}
@@ -1081,56 +631,6 @@ func processLocals(l log.Logger, parser *ParsingContext, opts *options.Terragrun
 	parser.Locals = &localsAsCtyVal
 
 	return nil
-}
-
-// listStackFiles searches for stack files in the specified directory.
-//
-// The function walks through the given directory to find files that match the
-// default stack file name. It optionally follows symbolic links based on the
-// provided Terragrunt options.
-func listStackFiles(l log.Logger, opts *options.TerragruntOptions, dir string) ([]string, error) {
-	walkWithSymlinks := opts.Experiments.Evaluate(experiment.Symlinks)
-
-	walkFunc := filepath.WalkDir
-	if walkWithSymlinks {
-		walkFunc = util.WalkDirWithSymlinks
-	}
-
-	l.Debugf("Searching for stack files in %s", dir)
-
-	var stackFiles []string
-
-	// find all defaultStackFile files
-	if err := walkFunc(dir, func(path string, d fs.DirEntry, err error) error {
-		if err != nil {
-			l.Warnf("Error accessing path %s: %w", path, err)
-			return nil
-		}
-
-		if d.IsDir() {
-			return nil
-		}
-
-		// skip files in Terragrunt cache directory
-		if strings.Contains(path, string(os.PathSeparator)+util.TerragruntCacheDir+string(os.PathSeparator)) ||
-			filepath.Base(path) == util.TerragruntCacheDir {
-			return filepath.SkipDir
-		}
-
-		if len(path) >= generationMaxPath {
-			return errors.Errorf("Cycle detected: maximum path length (%d) exceeded at %s", generationMaxPath, path)
-		}
-
-		if strings.HasSuffix(path, defaultStackFile) {
-			l.Debugf("Found stack file %s", path)
-			stackFiles = append(stackFiles, path)
-		}
-		return nil
-	}); err != nil {
-		return nil, err
-	}
-
-	return stackFiles, nil
 }
 
 // validateTargetDir target destination directory.
@@ -1194,8 +694,8 @@ func CleanStacks(_ context.Context, l log.Logger, opts *options.TerragruntOption
 	return errs.ErrorOrNil()
 }
 
-// getUnitDir returns the directory path for a unit based on its no_dot_terragrunt_stack setting.
-func getUnitDir(dir string, unit *Unit) string {
+// GetUnitDir returns the directory path for a unit based on its no_dot_terragrunt_stack setting.
+func GetUnitDir(dir string, unit *Unit) string {
 	if unit.NoStack != nil && *unit.NoStack {
 		return filepath.Join(dir, unit.Path)
 	}
